feat: Added queries for ReviewDao

+++ services-interfaces/src/main/java/ar/edu/itba/paw/services/ReviewService.java
@@ -2,6 +2,6 @@ package ar.edu.itba.paw.services;
 
 import ar.edu.itba.paw.models.Review;
 
-public interface ReviewService extends BaseService<Integer, Review> {
+public interface ReviewService extends BaseService<Long, Review> {
 
 }
